+++ src/views/HomeView.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+
+const HomeView = () => {
+
+
+    return (
+        <div>
+            <h1>Hello, Welcome to this Template</h1>
+            <p>
+                You can get more information in this <a target="_blank" href="https://github.com/kaito002/react-microfrontend-container-ts">link</a>, feel free 
+                to use this project as you want to.
+            </p>
+        </div>
+    )
+}
+
+export default HomeView;