+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.0
+* Using of React Router DOM
+* Accept Enviroment Files
+* Allows Server Side Rendering
+
 # 0.1.0
 * Initial Project
 * Basic Build